Updated CollisionSystem with threejs API and added to game loop.
For the CollisionSystem, the code to do AABB calculations were removed
because threeJS's Box3 API provides that functionality. The
CollisionSystem was also given an update loop and started in the game.
For these to work, a hitbox using Box3 was made and is recalculated on
every frame, but the hitbox is too big right now.