Added a basic set of ECS classes and changed the player object to it.
The basic ECS classes include Component and Entity, which many of the
components of the game object are now switched to such as input and
hitbox, etc. But the flaw is that the current logic only takes the name
of the component passed in, not the name of the interface.