Filled out the Leveling System and fixed level up calculations.
The Leveling System handles all data and logic on level up. It uses the
player's current level stats instead of the current play stats, which
represent the current state of the player. The game levels are
initialized with the play stats which are a copy of level stats, except
it is allowed to be updated while playing a level.